pyproject.toml
[project]
name = "ecommerce-async-tests"
version = "0.1.0"
description = "демонстрация асинхронных API-тестов"
authors = [
    {name = "Pavel",email = "savchukpa@icloud.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "allure-pytest (>=2.15.0,<3.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "faker (>=37.12.0,<38.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "validators (>=0.35.0,<0.36.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
